Support for multiple cmds
