+++ cmd/controller/cert_controller.go
@@ -19,13 +19,14 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
-package controller
+package main
 
 import (
 	"os"
 	"time"
 
 	"github.com/mariadb-operator/mariadb-operator/controller"
+	"github.com/mariadb-operator/mariadb-operator/pkg/log"
 	"github.com/spf13/cobra"
 	"k8s.io/apimachinery/pkg/types"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -42,12 +43,33 @@ var (
 	requeueDuration                               time.Duration
 )
 
+func init() {
+	rootCmd.AddCommand(certControllerCmd)
+	certControllerCmd.Flags().StringVar(&caSecretName, "ca-secret-name", "mariadb-operator-webhook-ca",
+		"Secret to store CA certificate for webhook")
+	certControllerCmd.Flags().StringVar(&caSecretNamespace, "ca-secret-namespace", "default",
+		"Namespace of the Secret to store the CA certificate for webhook")
+	certControllerCmd.Flags().StringVar(&caCommonName, "ca-common-name", "mariadb-operator", "CA certificate common name")
+	certControllerCmd.Flags().DurationVar(&caValidity, "ca-validity", 4*365*24*time.Hour, "CA certificate validity")
+	certControllerCmd.Flags().StringVar(&certSecretName, "cert-secret-name", "mariadb-operator-webhook-cert",
+		"Secret to store the certificate for webhook")
+	certControllerCmd.Flags().StringVar(&certSecretNamespace, "cert-secret-namespace", "default",
+		"Namespace of the Secret to store the certificate for webhook")
+	certControllerCmd.Flags().DurationVar(&certValidity, "cert-validity", 365*24*time.Hour, "Certificate validity")
+	certControllerCmd.Flags().DurationVar(&lookaheadValidity, "lookahead-validity", 90*24*time.Hour,
+		"Lookahead validity used to determine whether a certificate is valid or not")
+	certControllerCmd.Flags().StringVar(&serviceName, "service-name", "mariadb-operator-webhook", "Webhook service name")
+	certControllerCmd.Flags().StringVar(&serviceNamespace, "service-namespace", "default", "Webhook service namespace")
+	certControllerCmd.Flags().DurationVar(&requeueDuration, "requeue-duration", time.Minute*5,
+		"Time duration between reconciling webhook config for new certs")
+}
+
 var certControllerCmd = &cobra.Command{
 	Use:   "cert-controller",
 	Short: "MariaDB operator certificate controller.",
 	Long:  `Issues and injects certificates for validation and mutation webhooks.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		setupLogger()
+		log.SetupLogger(logLevel, logTimeEncoder, logDev)
 
 		mgr, err := ctrl.NewManager(ctrl.GetConfigOrDie(), ctrl.Options{
 			Scheme: scheme,
@@ -103,25 +125,4 @@ var certControllerCmd = &cobra.Command{
 			os.Exit(1)
 		}
 	},
-}
-
-func init() {
-	rootCmd.AddCommand(certControllerCmd)
-	certControllerCmd.Flags().StringVar(&caSecretName, "ca-secret-name", "mariadb-operator-webhook-ca",
-		"Secret to store CA certificate for webhook")
-	certControllerCmd.Flags().StringVar(&caSecretNamespace, "ca-secret-namespace", "default",
-		"Namespace of the Secret to store the CA certificate for webhook")
-	certControllerCmd.Flags().StringVar(&caCommonName, "ca-common-name", "mariadb-operator", "CA certificate common name")
-	certControllerCmd.Flags().DurationVar(&caValidity, "ca-validity", 4*365*24*time.Hour, "CA certificate validity")
-	certControllerCmd.Flags().StringVar(&certSecretName, "cert-secret-name", "mariadb-operator-webhook-cert",
-		"Secret to store the certificate for webhook")
-	certControllerCmd.Flags().StringVar(&certSecretNamespace, "cert-secret-namespace", "default",
-		"Namespace of the Secret to store the certificate for webhook")
-	certControllerCmd.Flags().DurationVar(&certValidity, "cert-validity", 365*24*time.Hour, "Certificate validity")
-	certControllerCmd.Flags().DurationVar(&lookaheadValidity, "lookahead-validity", 90*24*time.Hour,
-		"Lookahead validity used to determine whether a certificate is valid or not")
-	certControllerCmd.Flags().StringVar(&serviceName, "service-name", "mariadb-operator-webhook", "Webhook service name")
-	certControllerCmd.Flags().StringVar(&serviceNamespace, "service-namespace", "default", "Webhook service namespace")
-	certControllerCmd.Flags().DurationVar(&requeueDuration, "requeue-duration", time.Minute*5,
-		"Time duration between reconciling webhook config for new certs")
 }
